#8 -Step 1: npm run format
Running `npm run format` produces this diff for me locally. If we want GitHub actions to keep this consistent we need a starting point.

Potentially this was formatted with a different configuration than what is present in the repo? If so, can change the configuration and re-apply.

For context, this PR represents no personal opinion of mine beyond the project reliably formatting itself.